Added rudimentary login functionality and fixed file structure for frontend React
