+++ cst438-project2/src/components/Login/Login.js
@@ -1,0 +1,68 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import '../css/Login.css';
+
+// async function loginUser(credentials) {
+//     return fetch('http://localhost:8080/login', {
+//       method: 'POST',
+//       headers: {
+//         'Content-Type': 'application/json'
+//       },
+//       body: JSON.stringify(credentials)
+//     })
+//       .then(data => data.json())
+// }
+
+// export default function Login({ setToken }) {
+export default function Login({ setToken }) {
+
+  const [username, setUserName] = useState();
+  const [password, setPassword] = useState();
+
+  // const handleSubmit = async e => {
+  //   e.preventDefault();
+  //   const token = await loginUser({
+  //     username,
+  //     password
+  //   });
+  //   setToken(token);
+  // }
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    let hardcodedCred = {
+      username: 'admin',
+      password: 'admin'
+    }
+    if ((username === hardcodedCred.username) && (password === hardcodedCred.password)) {
+      //combination is good. Log them in.
+      //this token can be anything. You can use random.org to generate a random string;
+      const token = '123456abcdef';
+      setToken(token);
+    } else {
+      //bad combination
+      alert('wrong email or password combination');
+    }
+  }
+  return(
+    <div className="login-wrapper">
+      <h1>Please Log In</h1>
+      <form onSubmit={handleSubmit}>
+        <label>
+          <p>Username</p>
+          <input type="text" onChange={e => setUserName(e.target.value)}/>
+        </label>
+        <label>
+          <p>Password</p>
+          <input type="password" onChange={e => setPassword(e.target.value)}/>
+        </label>
+        <div>
+          <button type="submit">Submit</button>
+        </div>
+      </form>
+    </div>
+  )
+}
+
+Login.propTypes = {
+    setToken: PropTypes.func.isRequired
+}
